backend/requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pillow==11.0.0
psycopg2-binary==2.9.9
PyMySQL==1.1.1
Werkzeug==3.1.3
zipp==3.21.0
